backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
pytz>=2023.3
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
python-multipart>=0.0.9
typer>=0.9.0
mercadopago==2.2.1
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
twilio==8.10.0
resend==2.0.0
pywebpush>=2.0.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-groq>=0.2.0
groq>=0.4.1,<1
openai>=1.0.0
langfuse>=3.0.0
